Update `xportr_type()` to drop column attribute 'class' only
In general, setting the type for each column should retain each column's
attributes, such as label, width, and sas.format. These are set by other
xportr functions. However, in cases where a column in the input dataset
has a class already, such as a Date or a factor, this attribute should
be dropped from the column. If it is not, it causes an error